feat: add onboarding landing page and dedicated discovery feed route
- Create `Onboarding` component explaining PersonaComposer core concepts (OCEAN, Focus Groups, Challenge Mode).
- Refactor discovery feed logic into a standalone `DiscoveryFeed` component.
- Implement conditional rendering on home page: Onboarding for guests, Feed for authenticated users.
- Add dedicated `/feed` route for guest browsing and improved navigation.
- Add descriptive blurb to the discovery feed explaining its purpose.
- Ensure responsive design and accessibility.
